Re-worked some aspects so it'll massively reduce the load time impact & the potential memory that is attempted to be allocated even if the plug-ins then never get used for playback during the running instance Disabled a bunch of debugging / logging not needed Fixed some metadata requests being incorrectly handled

+++ src/in_2sf/desmume/emufile.cpp
@@ -27,7 +27,7 @@ THE SOFTWARE.
 size_t EMUFILE_MEMORY::_fread(void *ptr, size_t bytes)
 {
 	uint32_t remain = this->len - this->pos;
-	uint32_t todo = std::min<uint32_t>(remain, bytes);
+	uint32_t todo = std::min<uint32_t>(remain, static_cast<uint32_t>(bytes));
 	if (!len)
 	{
 		this->failbit = true;
